+++ mycodo/mycodo_flask/forms/forms_lcd.py
@@ -68,7 +68,7 @@ class LCDMod(FlaskForm):
 class LCDModDisplay(FlaskForm):
     lcd_id = StringField('LCD ID', widget=widgets.HiddenInput())
     lcd_data_id = StringField('LCD Data ID', widget=widgets.HiddenInput())
-    line_1_display = StringField(lazy_gettext('Line 1'))
+    line_1_display = StringField("{}: 1".format(TRANSLATIONS['line']['title']))
     line_1_max_age = IntegerField(
         TRANSLATIONS['max_age']['title'],
         validators=[validators.NumberRange(min=1, max=999999999)],
@@ -79,7 +79,7 @@ class LCDModDisplay(FlaskForm):
         validators=[validators.NumberRange(min=0)],
         widget=NumberInput()
     )
-    line_2_display = StringField(lazy_gettext('Line 2'))
+    line_2_display = StringField("{}: 2".format(TRANSLATIONS['line']['title']))
     line_2_max_age = IntegerField(
         TRANSLATIONS['max_age']['title'],
         validators=[validators.NumberRange(min=1, max=999999999)],
@@ -90,7 +90,7 @@ class LCDModDisplay(FlaskForm):
         validators=[validators.NumberRange(min=0)],
         widget=NumberInput()
     )
-    line_3_display = StringField(lazy_gettext('Line 3'))
+    line_3_display = StringField("{}: 3".format(TRANSLATIONS['line']['title']))
     line_3_max_age = IntegerField(
         TRANSLATIONS['max_age']['title'],
         validators=[validators.NumberRange(min=1, max=999999999)],
@@ -101,7 +101,7 @@ class LCDModDisplay(FlaskForm):
         validators=[validators.NumberRange(min=0)],
         widget=NumberInput()
     )
-    line_4_display = StringField(lazy_gettext('Line 4'))
+    line_4_display = StringField("{}: 4".format(TRANSLATIONS['line']['title']))
     line_4_max_age = IntegerField(
         TRANSLATIONS['max_age']['title'],
         validators=[validators.NumberRange(min=1, max=999999999)],
@@ -112,5 +112,50 @@ class LCDModDisplay(FlaskForm):
         validators=[validators.NumberRange(min=0)],
         widget=NumberInput()
     )
+    line_5_display = StringField("{}: 5".format(TRANSLATIONS['line']['title']))
+    line_5_max_age = IntegerField(
+        TRANSLATIONS['max_age']['title'],
+        validators=[validators.NumberRange(min=1, max=999999999)],
+        widget=NumberInput()
+    )
+    line_5_decimal_places = IntegerField(
+        lazy_gettext('Decimal Places'),
+        validators=[validators.NumberRange(min=0)],
+        widget=NumberInput()
+    )
+    line_6_display = StringField("{}: 6".format(TRANSLATIONS['line']['title']))
+    line_6_max_age = IntegerField(
+        TRANSLATIONS['max_age']['title'],
+        validators=[validators.NumberRange(min=1, max=999999999)],
+        widget=NumberInput()
+    )
+    line_6_decimal_places = IntegerField(
+        lazy_gettext('Decimal Places'),
+        validators=[validators.NumberRange(min=0)],
+        widget=NumberInput()
+    )
+    line_7_display = StringField("{}: 7".format(TRANSLATIONS['line']['title']))
+    line_7_max_age = IntegerField(
+        TRANSLATIONS['max_age']['title'],
+        validators=[validators.NumberRange(min=1, max=999999999)],
+        widget=NumberInput()
+    )
+    line_7_decimal_places = IntegerField(
+        lazy_gettext('Decimal Places'),
+        validators=[validators.NumberRange(min=0)],
+        widget=NumberInput()
+    )
+    line_8_display = StringField("{}: 8".format(TRANSLATIONS['line']['title']))
+    line_8_max_age = IntegerField(
+        TRANSLATIONS['max_age']['title'],
+        validators=[validators.NumberRange(min=1, max=999999999)],
+        widget=NumberInput()
+    )
+    line_8_decimal_places = IntegerField(
+        lazy_gettext('Decimal Places'),
+        validators=[validators.NumberRange(min=0)],
+        widget=NumberInput()
+    )
+
     save_display = SubmitField(TRANSLATIONS['save']['title'])
     delete_display = SubmitField(TRANSLATIONS['delete']['title'])